Revert code since multiple values and optional is already handled in typer cli
